analogous_to_modules
