+++ examples/mlp.py
@@ -5,14 +5,14 @@ Myia is still a work in progress, and this example may change in the future.
 
 
 import time
-from dataclasses import dataclass
 
 import numpy
 from numpy.random import RandomState
 
-from myia import ArithmeticData, myia, value_and_grad
+from myia import myia, value_and_grad
 from myia.api import to_device
 from myia.debug import traceback  # noqa
+from myia.modules import Linear, Sequential, Tanh
 
 ###########
 # Options #
@@ -79,49 +79,6 @@ def mlp_parameters(*layer_sizes, seed=90909):
         b = param(R, 1, o)
         parameters.append((W, b))
     return parameters
-
-
-#########
-# Model #
-#########
-
-
-# We generate a MLP model with some arbitrary number of layers and tanh
-# activations.
-
-
-@dataclass(frozen=True)
-class Linear(ArithmeticData):
-    """Linear layer."""
-
-    W: "Weights array"
-    b: "Biases vector"
-
-    def apply(self, input):
-        """Apply the layer."""
-        return input @ self.W + self.b
-
-
-@dataclass(frozen=True)
-class Tanh(ArithmeticData):
-    """Tanh layer."""
-
-    def apply(self, input):
-        """Apply the layer."""
-        return numpy.tanh(input)
-
-
-@dataclass(frozen=True)
-class Sequential(ArithmeticData):
-    """Sequential layer, applies all sub-layers in order."""
-
-    layers: "Tuple of layers"
-
-    def apply(self, x):
-        """Apply the layer."""
-        for layer in self.layers:
-            x = layer.apply(x)
-        return x
 
 
 def cost(model, x, target):
